+++ src/test/java/io/github/thebusybiscuit/slimefun4/tests/listeners/TestBackpackListener.java
@@ -6,6 +6,7 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
+import org.bukkit.NamespacedKey;
 import org.bukkit.entity.Item;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.ClickType;
@@ -28,13 +29,15 @@ import org.mockito.Mockito;
 import be.seeseemelk.mockbukkit.MockBukkit;
 import be.seeseemelk.mockbukkit.ServerMock;
 import io.github.thebusybiscuit.cscorelib2.item.CustomItem;
-import io.github.thebusybiscuit.slimefun4.api.items.ItemState;
 import io.github.thebusybiscuit.slimefun4.api.player.PlayerBackpack;
 import io.github.thebusybiscuit.slimefun4.api.player.PlayerProfile;
 import io.github.thebusybiscuit.slimefun4.implementation.items.backpacks.SlimefunBackpack;
 import io.github.thebusybiscuit.slimefun4.implementation.listeners.BackpackListener;
 import io.github.thebusybiscuit.slimefun4.mocks.TestUtilities;
 import me.mrCookieSlime.Slimefun.SlimefunPlugin;
+import me.mrCookieSlime.Slimefun.Lists.RecipeType;
+import me.mrCookieSlime.Slimefun.Objects.Category;
+import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
 public class TestBackpackListener {
 
@@ -70,19 +73,16 @@ public class TestBackpackListener {
         return ref.get();
     }
 
-    private PlayerBackpack openMockBackpack(Player player, int size) throws InterruptedException {
-        ItemStack item = new CustomItem(Material.CHEST, "&4Mock Backpack", "", "&7Size: &e" + BACKPACK_SIZE, "&7ID: <ID>", "", "&7&eRight Click&7 to open");
+    private PlayerBackpack openMockBackpack(Player player, String id, int size) throws InterruptedException {
+        SlimefunItemStack item = new SlimefunItemStack(id, Material.CHEST, "&4Mock Backpack", "", "&7Size: &e" + BACKPACK_SIZE, "&7ID: <ID>", "", "&7&eRight Click&7 to open");
         PlayerProfile profile = TestUtilities.awaitProfile(player);
 
         PlayerBackpack backpack = profile.createBackpack(size);
         listener.setBackpackId(player, item, 2, backpack.getId());
 
-        SlimefunBackpack slimefunBackpack = Mockito.mock(SlimefunBackpack.class);
-        Mockito.when(slimefunBackpack.getSize()).thenReturn(size);
-
-        // This will make hasUnlocked() immediately pass
-        Mockito.when(slimefunBackpack.getState()).thenReturn(ItemState.VANILLA_FALLBACK);
-        SlimefunPlugin.getRegistry().getEnabledSlimefunItems().add(slimefunBackpack);
+        Category category = new Category(new NamespacedKey(plugin, "test_backpacks"), new CustomItem(Material.CHEST, "&4Test Backpacks"));
+        SlimefunBackpack slimefunBackpack = new SlimefunBackpack(size, category, item, RecipeType.NULL, new ItemStack[9]);
+        slimefunBackpack.register(plugin);
 
         listener.openBackpack(player, item, slimefunBackpack);
         return backpack;
@@ -118,7 +118,7 @@ public class TestBackpackListener {
     @Test
     public void testOpenBackpack() throws InterruptedException {
         Player player = server.addPlayer();
-        PlayerBackpack backpack = openMockBackpack(player, 27);
+        PlayerBackpack backpack = openMockBackpack(player, "TEST_OPEN_BACKPACK", 27);
         InventoryView view = player.getOpenInventory();
         Assertions.assertEquals(backpack.getInventory(), view.getTopInventory());
     }
@@ -126,7 +126,7 @@ public class TestBackpackListener {
     @Test
     public void testCloseBackpack() throws InterruptedException {
         Player player = server.addPlayer();
-        PlayerBackpack backpack = openMockBackpack(player, 27);
+        PlayerBackpack backpack = openMockBackpack(player, "TEST_CLOSE_BACKPACK", 27);
         listener.onClose(new InventoryCloseEvent(player.getOpenInventory()));
 
         Assertions.assertTrue(backpack.getOwner().isDirty());
@@ -135,7 +135,7 @@ public class TestBackpackListener {
     @Test
     public void testBackpackDropNormalItem() throws InterruptedException {
         Player player = server.addPlayer();
-        openMockBackpack(player, 27);
+        openMockBackpack(player, "DROP_NORMAL_ITEM_BACKPACK_TEST", 27);
 
         Item item = Mockito.mock(Item.class);
         Mockito.when(item.getItemStack()).thenReturn(new ItemStack(Material.SUGAR_CANE));
@@ -145,9 +145,9 @@ public class TestBackpackListener {
         Assertions.assertFalse(event.isCancelled());
     }
 
-    private boolean isAllowed(ItemStack item) throws InterruptedException {
+    private boolean isAllowed(String id, ItemStack item) throws InterruptedException {
         Player player = server.addPlayer();
-        Inventory inv = openMockBackpack(player, 9).getInventory();
+        Inventory inv = openMockBackpack(player, id, 9).getInventory();
 
         int slot = 7;
         inv.setItem(slot, item);
@@ -159,20 +159,20 @@ public class TestBackpackListener {
     @ParameterizedTest
     @EnumSource(value = Material.class, names = { "AIR", "DIAMOND", "STONE" })
     public void areItemsAllowed(Material type) throws InterruptedException {
-        Assertions.assertTrue(isAllowed(new ItemStack(type)));
+        Assertions.assertTrue(isAllowed("BACKPACK_ALLOWANCE_" + type.name(), new ItemStack(type)));
     }
 
     @ParameterizedTest
     @EnumSource(value = Material.class, names = { "SHULKER_BOX", "RED_SHULKER_BOX", "BLUE_SHULKER_BOX", "BLACK_SHULKER_BOX" })
     public void areShulkerBoxesAllowed(Material type) throws InterruptedException {
-        Assertions.assertFalse(isAllowed(new ItemStack(type)));
+        Assertions.assertFalse(isAllowed("BACKPACK_ALLOWANCE_" + type.name(), new ItemStack(type)));
     }
 
     @ParameterizedTest
     @EnumSource(value = Material.class, names = { "AIR", "SHULKER_BOX" })
     public void testHotbarKey(Material type) throws InterruptedException {
         Player player = server.addPlayer();
-        openMockBackpack(player, 9);
+        openMockBackpack(player, "BACKPACK_HOTBAR_" + type.name(), 9);
 
         int slot = 7;
         player.getInventory().setItem(slot, new ItemStack(type));