+++ CHANGELOG.md
@@ -24,11 +24,16 @@
 * Added Dried Kelp Blocks recipe to the Electric Press
 * Added Bone Blocks recipe to the Electric Press
 * Added thai translations
+* Added Industrial Miner
+* Added Advanced Industrial Miner
 * Added Cocoa Organic Food
 * Added Cocoa Fertilizer
 
 #### Changes
 * Fixed a few memory leaks
+* Removed Digital Miner
+* Removed Advanced Digital Miner
+* Dried Kelp Blocks can now be used in the Coal Generator
 * Crafting Organic Food/Fertilizer yields more output now
 * Organic Food (Melon) now uses Melon Slices instead of Melon blocks
 
@@ -39,6 +44,7 @@
 * Fixed Nuclear Reactors accepting Lava as coolant
 
 ## Release Candidate 12 (27 May 2020)
+https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#12
 
 #### Additions
 * Added Ukrainian translations
@@ -83,6 +89,7 @@
 * Fixed #1935
 
 ## Release Candidate 11 (25 Apr 2020)
+https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#11
 
 #### Additions
 * Added GEOResourceGenerationEvent